feat: 22: refactor styling
styles.js: This file has been added and all styling has been moved to
this file.
