+++ Page1.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { Button, ScrollView, StyleSheet, Text, TextInput, View } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
 import addToFireStore from './firestore.js'
+import styles from './styles.js'
 const Page1 = () => {
     const navigation = useNavigation();
     //const [text, onChangeText] = React.useState('Useless Text');
@@ -15,27 +16,5 @@ const Page1 = () => {
     </View>
     );
   };
-  
-  const styles = StyleSheet.create({
-    container: {
-      flex: 1,
-      backgroundColor: '#fff',
-      alignItems: 'center',
-      justifyContent: 'center',
-    },
-    container2: {
-      flex: 1,
-    },
-    input: {
-      height: 40,
-      width: 120,
-      margin: 12,
-      borderWidth: 1,
-      padding: 10,
-    },
-    map:{
-        width: '100%',
-        height: '100%',
-    }
-  });
+
 export default Page1
