+++ Page2.js
@@ -2,21 +2,12 @@ import React from 'react';
 import { Button, ScrollView, StyleSheet, Text, TextInput, View } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
 import MapView from 'react-native-maps';
-
+import styles from './styles'
 export default function Map() {
     return (
-      <View style={styles.container}>
+      <View style={styles.map_container}>
         <MapView style={styles.map} />
       </View>
     );
   }
   
-  const styles = StyleSheet.create({
-    container: {
-      flex: 1,
-    },
-    map: {
-      width: '100%',
-      height: '100%',
-    },
-  });
