+++ styles.js
@@ -1,0 +1,30 @@
+import { StyleSheet } from 'react-native';
+
+
+const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+      backgroundColor: '#fff',
+      alignItems: 'center',
+      justifyContent: 'center',
+    },
+    input: {
+      height: 40,
+      width: 120,
+      margin: 12,
+      borderWidth: 1,
+      padding: 10,
+    },
+    logo: {
+        width: 200,
+        height: 200,
+      },
+      map_container: {
+        flex: 1,
+      },
+      map:{
+          width: '100%',
+          height: '100%',
+      }
+  });
+export default styles
